hello_tools/pyproject.toml
[project]
name = "hello-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.3.3",
    "tavily-python>=0.7.12",
    "requests>=2.31.0",
]
